Prod release 4/8/21

+++ Source/OregonTilth.Models/DataTransferObjects/Workbook/Forms/TimeStudy/TimeStudyUpsertDto.cs
@@ -7,8 +7,8 @@
         public int? HarvestPostHarvestStandardTimeID { get; set; }
         public int? TransplantProductionStandardTimeID { get; set; }
         public string Notes { get; set; }
-        public int Units { get; set; }
-        public int Duration { get; set; }
+        public decimal Units { get; set; }
+        public decimal Duration { get; set; }
         public int WorkbookID { get; set; }
     }
 }
